[DTensor] separate out test xfails
[ghstack-poisoned]
