+++ test/distributed/tensor/test_dtensor_ops.py
@@ -25,12 +25,18 @@ from torch.testing._internal.common_device_type import (
     ops,
 )
 from torch.testing._internal.common_methods_invocations import DecorateInfo, op_db
-from torch.testing._internal.common_ops_unbacked import ops_dde_xfail, ops_unbacked_skip
 from torch.testing._internal.common_utils import run_tests, suppress_warnings, TestCase
 from torch.testing._internal.distributed._tensor.common_dtensor import (
     DTensorConverter,
     DTensorOpTestBase,
     validate_sharding_rule_sample,
+)
+from torch.testing._internal.distributed._tensor.dtensor_xfails import (
+    dtensor_compiled_xfails,
+    dtensor_xfails,
+    multi_threaded_xfails,
+    no_strategy_xfails,
+    unbacked_dtensor_xfails,
 )
 from torch.utils import _pytree as pytree
 from torch.utils._debug_mode import _OpCall, DebugMode
@@ -44,15 +50,6 @@ common_ops.L = 24
 common_ops.M = 12
 common_ops.S = 4
 common_ops.XS = 2
-
-
-# Copied from functorch
-def xfail(op_name, variant_name="", *, device_type=None, dtypes=None):
-    return (op_name, variant_name, device_type, dtypes, True)
-
-
-def skip(op_name, variant_name="", *, device_type=None, dtypes=None):
-    return (op_name, variant_name, device_type, dtypes, False)
 
 
 def skipOps(op_db, test_case_name, base_test_name, to_skip):
@@ -121,445 +118,6 @@ def repurpose_ops(op_db, base_test_name, derived_test_name):
         repurposed_ops.append(opinfo_copy)
     return repurposed_ops
 
-
-# Re-generate this failed list, turn on dry_run of the below func
-# check_dtensor_func(self, test, op, dry_run=True), then run sth
-# like python test/distributed/tensor/test_dtensor_ops.py > failed.expect
-dtensor_fails = {
-    # these sometimes pass and sometimes fail
-    # we need to remove many of them from list once op
-    # get full support with varying sharding specs
-    xfail("__getitem__"),
-    xfail("_chunk_cat"),
-    xfail("arange"),
-    xfail("as_strided"),
-    xfail("as_strided", "partial_views"),
-    xfail("bernoulli"),
-    xfail("block_diag"),
-    xfail("broadcast_shapes"),
-    xfail("cartesian_prod"),
-    xfail("cholesky_solve"),
-    xfail("eye"),
-    xfail("flatten"),
-    xfail("full"),
-    xfail("index_put"),
-    xfail("kron"),
-    xfail("linalg.matrix_power"),
-    xfail("linspace"),
-    xfail("linspace", "tensor_overload"),
-    xfail("logdet"),
-    xfail("logspace"),
-    xfail("logspace", "tensor_overload"),
-    xfail("masked.logsumexp"),
-    xfail("masked.median"),
-    xfail("masked_scatter"),
-    xfail("masked_select"),
-    xfail("nanquantile"),
-    xfail("transpose"),
-    xfail("nn.functional.adaptive_avg_pool1d"),
-    xfail("nn.functional.adaptive_avg_pool3d"),
-    xfail("nn.functional.alpha_dropout"),
-    xfail("nn.functional.conv1d"),
-    xfail("nn.functional.conv2d"),
-    xfail("nn.functional.conv3d"),
-    xfail("nn.functional.conv_transpose1d"),
-    xfail("nn.functional.conv_transpose2d"),
-    xfail("nn.functional.conv_transpose3d"),
-    xfail("nn.functional.dropout"),
-    xfail("nn.functional.fractional_max_pool2d"),
-    xfail("nn.functional.fractional_max_pool3d"),
-    xfail("nn.functional.interpolate", "area"),
-    xfail("nn.functional.interpolate", "nearest"),
-    xfail("nn.functional.interpolate", "nearest-exact"),
-    xfail("nn.functional.pad", "reflect"),
-    xfail("nn.functional.pad", "replicate"),
-    xfail("nn.functional.pad", "replicate_negative"),
-    xfail("nn.functional.unfold"),
-    xfail("normal"),
-    xfail("normal", "number_mean"),
-    xfail("normal", "in_place"),
-    xfail("ones"),
-    xfail("pca_lowrank"),
-    xfail("quantile"),
-    xfail("rand_like"),
-    xfail("randint_like"),
-    xfail("randint"),
-    xfail("randn"),
-    xfail("randn_like"),
-    xfail("ravel"),
-    xfail("repeat_interleave"),
-    xfail("reshape"),
-    xfail("reshape_as"),
-    xfail("roll"),
-    xfail("scalar_tensor"),
-    xfail("searchsorted"),
-    xfail("sparse.sampled_addmm"),
-    xfail("sparse.mm", "reduce"),
-    xfail("squeeze", "multiple"),
-    xfail("signal.windows.bartlett"),
-    xfail("signal.windows.blackman"),
-    xfail("signal.windows.cosine"),
-    xfail("signal.windows.exponential"),
-    xfail("signal.windows.gaussian"),
-    xfail("signal.windows.general_cosine"),
-    xfail("signal.windows.general_hamming"),
-    xfail("signal.windows.hamming"),
-    xfail("signal.windows.hann"),
-    xfail("signal.windows.nuttall"),
-    xfail("signal.windows.kaiser"),
-    xfail("stack"),
-    xfail("svd_lowrank"),
-    xfail("take_along_dim"),
-    xfail("tensor_split"),
-    xfail("unbind"),
-    xfail("uniform"),
-    xfail("unflatten"),
-    xfail("unsafe_chunk"),
-    xfail("_unsafe_masked_index"),
-    xfail("_unsafe_masked_index_put_accumulate"),
-    xfail("view"),
-    xfail("view_as"),
-    xfail("zeros"),
-    # /TODO(whc) debug/triage
-    # ops inside this might even fail without dtensor
-    # tests, as we rescale op db common test size factor (i.e. L, M, S)
-    # which triggered the original function run failures with input
-    # generation becomes wrong, we skip them for now but should enable later.
-    # TODO: need to clean this list and remove all cases
-    skip("argwhere"),
-    skip("cumprod"),
-    skip("__rmatmul__"),
-    skip("meshgrid", "list_of_tensors"),
-    skip("meshgrid", "variadic_tensors"),
-    skip("nn.functional.scaled_dot_product_attention"),
-    skip("nn.functional.softmin"),
-    skip("nn.functional.embedding"),
-    skip("nn.functional.embedding_bag"),
-    skip("nn.functional.feature_alpha_dropout", "with_train"),
-    skip("nn.functional.feature_alpha_dropout", "without_train"),
-    skip("nn.functional.hinge_embedding_loss"),
-    skip("nn.functional.cosine_embedding_loss"),
-    skip("fft.hfft"),
-    skip("fft.hfft2"),
-    skip("fft.hfft2"),
-    skip("fft.hfftn"),
-    skip("fft.ifftn"),
-    skip("fft.irfft"),
-    skip("istft"),
-    skip("isclose"),
-    skip("isreal"),
-    skip("matmul"),
-    skip("masked.mean"),
-    skip("masked.var"),
-    skip("masked.std"),
-    skip("masked.normalize"),
-    skip("prod"),
-    skip("_segment_reduce", "lengths"),
-    skip("_segment_reduce", "offsets"),
-    # TODO: fix the following ops
-    skip("squeeze"),
-    skip("empty"),
-    skip("empty_strided"),
-    skip("empty_like"),
-    skip("empty_permuted"),
-    skip("new_empty"),
-    skip("new_empty_strided"),
-}
-
-dtensor_multi_threaded_fails = {
-    xfail("full_like"),
-    xfail("nn.functional.dropout2d"),
-    xfail("nn.functional.dropout3d"),
-    xfail("nn.functional.huber_loss"),
-    xfail("masked.cumprod"),
-    skip("nn.functional.multi_head_attention_forward"),
-}
-
-# Ops that fail to compile with DTensor + torch.compile(fullgraph=True).
-# These are compile-time failures (as_strided, data-dependent, etc.),
-# NOT numeric correctness issues.
-dtensor_compiled_fails = {
-    xfail("argmax"),
-    xfail("argmin"),
-    xfail("atleast_1d"),
-    xfail("atleast_2d"),
-    xfail("atleast_3d"),
-    xfail("broadcast_tensors"),
-    xfail("broadcast_to"),
-    xfail("corrcoef"),
-    xfail("cov"),
-    xfail("diagonal"),
-    xfail("dsplit"),
-    xfail("equal"),
-    xfail("expand"),
-    xfail("expand_as"),
-    xfail("gather"),
-    xfail("hash_tensor"),
-    xfail("hsplit"),
-    xfail("index_select"),
-    xfail("item"),
-    xfail("linalg.diagonal"),
-    xfail("masked.argmax"),
-    xfail("masked.argmin"),
-    xfail("max", "reduction_with_dim"),
-    xfail("min", "reduction_with_dim"),
-    xfail("movedim"),
-    xfail("narrow"),
-    xfail("nn.functional.binary_cross_entropy"),
-    xfail("nn.functional.binary_cross_entropy_with_logits"),
-    xfail("nn.functional.gaussian_nll_loss"),
-    xfail("nn.functional.interpolate", "bicubic"),
-    xfail("nn.functional.interpolate", "bilinear"),
-    xfail("nn.functional.interpolate", "linear"),
-    xfail("nn.functional.interpolate", "trilinear"),
-    xfail("nn.functional.logsigmoid"),
-    xfail("nn.functional.margin_ranking_loss"),
-    xfail("nn.functional.multilabel_soft_margin_loss"),
-    xfail("nn.functional.upsample_bilinear"),
-    xfail("nonzero_static"),
-    xfail("permute"),
-    xfail("scatter"),
-    xfail("scatter_add"),
-    xfail("select"),
-    xfail("slice"),
-    xfail("t"),
-    xfail("transpose_copy"),
-    xfail("unsqueeze"),
-    xfail("vsplit"),
-}
-
-# Ops that compile successfully but fail numeric checks in eager DTensor tests.
-# These are excluded from TestCompiledDTensorOps skip list since we don't check numerics.
-dtensor_numeric_only_fails = {
-    xfail("arange"),
-    xfail("broadcast_shapes"),
-    xfail("eye"),
-    xfail("full"),
-    xfail("full_like"),
-    xfail("linspace"),
-    xfail("logspace"),
-    xfail("narrow_copy"),
-    xfail("nn.functional.glu"),
-    xfail("nn.functional.hardshrink"),
-    xfail("nn.functional.huber_loss"),
-    xfail("nn.functional.smooth_l1_loss"),
-    xfail("nn.functional.softshrink"),
-    xfail("ones"),
-    xfail("randint"),
-    xfail("randn"),
-    xfail("scalar_tensor"),
-    xfail("signal.windows.bartlett"),
-    xfail("signal.windows.blackman"),
-    xfail("signal.windows.cosine"),
-    xfail("signal.windows.exponential"),
-    xfail("signal.windows.gaussian"),
-    xfail("signal.windows.general_cosine"),
-    xfail("signal.windows.general_hamming"),
-    xfail("signal.windows.hamming"),
-    xfail("signal.windows.hann"),
-    xfail("signal.windows.kaiser"),
-    xfail("signal.windows.nuttall"),
-    xfail("sparse.mm", "reduce"),
-    xfail("squeeze_copy"),
-    xfail("stack"),
-    xfail("unsafe_chunk"),
-    xfail("zeros"),
-}
-
-# Ops in dtensor_fails that have no sharding strategy (NotImplementedError).
-# These will error during sharding propagation and affect unbacked tests too.
-dtensor_fails_no_strategy = {
-    xfail("_batch_norm_with_update"),
-    xfail("_native_batch_norm_legit"),
-    xfail("_upsample_bilinear2d_aa"),
-    xfail("addbmm"),
-    xfail("allclose"),
-    xfail("as_strided_copy"),
-    xfail("as_strided_scatter"),
-    xfail("cauchy"),
-    xfail("cdist"),
-    xfail("cholesky"),
-    xfail("cholesky_inverse"),
-    xfail("combinations"),
-    xfail("complex"),
-    xfail("cross"),
-    xfail("cummax"),
-    xfail("cummin"),
-    xfail("diagonal_scatter"),
-    xfail("exponential"),
-    xfail("fft.fft"),
-    xfail("fft.fft2"),
-    xfail("fft.fftn"),
-    xfail("fft.fftshift"),
-    xfail("fft.ifft"),
-    xfail("fft.ifft2"),
-    xfail("fft.ifftshift"),
-    xfail("fft.ihfft"),
-    xfail("fft.ihfft2"),
-    xfail("fft.ihfftn"),
-    xfail("fft.irfft2"),
-    xfail("fft.irfftn"),
-    xfail("fft.rfft"),
-    xfail("fft.rfft2"),
-    xfail("fft.rfftn"),
-    xfail("flip"),
-    xfail("fliplr"),
-    xfail("flipud"),
-    xfail("frexp"),
-    xfail("geometric"),
-    xfail("geqrf"),
-    xfail("grid_sampler_2d"),
-    xfail("histogram"),
-    xfail("histogramdd"),
-    xfail("index_add"),
-    xfail("index_copy"),
-    xfail("index_fill"),
-    xfail("index_reduce", "prod"),
-    xfail("index_reduce", "mean"),
-    xfail("index_reduce", "amax"),
-    xfail("index_reduce", "amin"),
-    xfail("isin"),
-    xfail("kthvalue"),
-    xfail("linalg.cholesky"),
-    xfail("linalg.cholesky_ex"),
-    xfail("linalg.cross"),
-    xfail("linalg.det"),
-    xfail("linalg.eig"),
-    xfail("linalg.eigvals"),
-    xfail("linalg.householder_product"),
-    xfail("linalg.inv"),
-    xfail("linalg.inv_ex"),
-    xfail("linalg.ldl_factor"),
-    xfail("linalg.ldl_factor_ex"),
-    xfail("linalg.ldl_solve"),
-    xfail("linalg.lstsq"),
-    xfail("linalg.lstsq", "grad_oriented"),
-    xfail("linalg.lu"),
-    xfail("linalg.lu_factor"),
-    xfail("linalg.lu_factor_ex"),
-    xfail("linalg.lu_solve"),
-    xfail("linalg.pinv"),
-    xfail("linalg.pinv", "hermitian"),
-    xfail("linalg.slogdet"),
-    xfail("linalg.solve"),
-    xfail("linalg.solve_ex"),
-    xfail("linalg.solve_triangular"),
-    xfail("linalg.tensorinv"),
-    xfail("linalg.tensorsolve"),
-    xfail("linalg.vander"),
-    xfail("log_normal"),
-    xfail("logcumsumexp"),
-    xfail("logsumexp"),
-    xfail("lu"),
-    xfail("lu_solve"),
-    xfail("lu_unpack"),
-    xfail("matrix_exp"),
-    xfail("max_pool2d_with_indices_backward"),
-    xfail("median"),
-    xfail("mode"),
-    xfail("multinomial"),
-    xfail("nanmean"),
-    xfail("nanmedian"),
-    xfail("nansum"),
-    xfail("narrow_copy"),
-    xfail("native_batch_norm"),
-    xfail("ne"),
-    xfail("nn.functional.adaptive_avg_pool2d"),
-    xfail("nn.functional.adaptive_max_pool1d"),
-    xfail("nn.functional.adaptive_max_pool2d"),
-    xfail("nn.functional.adaptive_max_pool3d"),
-    xfail("nn.functional.avg_pool1d"),
-    xfail("nn.functional.avg_pool2d"),
-    xfail("nn.functional.avg_pool3d"),
-    xfail("nn.functional.batch_norm"),
-    xfail("nn.functional.batch_norm", "without_cudnn"),
-    xfail("nn.functional.bilinear"),
-    xfail("nn.functional.cosine_similarity"),
-    xfail("nn.functional.ctc_loss"),
-    xfail("nn.functional.glu"),
-    xfail("nn.functional.grid_sample"),
-    xfail("nn.functional.group_norm"),
-    xfail("nn.functional.hardshrink"),
-    xfail("nn.functional.instance_norm"),
-    xfail("nn.functional.local_response_norm"),
-    xfail("nn.functional.max_pool1d"),
-    xfail("nn.functional.max_pool2d"),
-    xfail("nn.functional.max_pool3d"),
-    xfail("nn.functional.max_unpool1d"),
-    xfail("nn.functional.max_unpool1d", "grad"),
-    xfail("nn.functional.max_unpool2d"),
-    xfail("nn.functional.max_unpool2d", "grad"),
-    xfail("nn.functional.max_unpool3d"),
-    xfail("nn.functional.max_unpool3d", "grad"),
-    xfail("nn.functional.multi_margin_loss"),
-    xfail("nn.functional.multilabel_margin_loss"),
-    xfail("nn.functional.pdist"),
-    xfail("nn.functional.rrelu"),
-    xfail("nn.functional.threshold"),
-    xfail("nn.functional.upsample_nearest"),
-    xfail("nonzero"),
-    xfail("ormqr"),
-    xfail("pinverse"),
-    xfail("polar"),
-    xfail("put"),
-    xfail("renorm"),
-    xfail("resize_"),
-    xfail("resize_as_"),
-    xfail("rot90"),
-    xfail("scatter_reduce", "amax"),
-    xfail("scatter_reduce", "amin"),
-    xfail("scatter_reduce", "mean"),
-    xfail("scatter_reduce", "prod"),
-    xfail("scatter_reduce", "sum"),
-    xfail("select_scatter"),
-    xfail("special.airy_ai"),
-    xfail("special.bessel_j0"),
-    xfail("special.bessel_j1"),
-    xfail("special.bessel_y0"),
-    xfail("special.bessel_y1"),
-    xfail("special.chebyshev_polynomial_t"),
-    xfail("special.chebyshev_polynomial_u"),
-    xfail("special.chebyshev_polynomial_v"),
-    xfail("special.chebyshev_polynomial_w"),
-    xfail("special.entr"),
-    xfail("special.erfcx"),
-    xfail("special.hermite_polynomial_h"),
-    xfail("special.hermite_polynomial_he"),
-    xfail("special.i0e"),
-    xfail("special.i1"),
-    xfail("special.i1e"),
-    xfail("special.laguerre_polynomial_l"),
-    xfail("special.legendre_polynomial_p"),
-    xfail("special.log_ndtr"),
-    xfail("special.modified_bessel_i0"),
-    xfail("special.modified_bessel_i1"),
-    xfail("special.modified_bessel_k0"),
-    xfail("special.modified_bessel_k1"),
-    xfail("special.ndtri"),
-    xfail("special.scaled_modified_bessel_k0"),
-    xfail("special.scaled_modified_bessel_k1"),
-    xfail("special.shifted_chebyshev_polynomial_t"),
-    xfail("special.shifted_chebyshev_polynomial_u"),
-    xfail("special.shifted_chebyshev_polynomial_v"),
-    xfail("special.shifted_chebyshev_polynomial_w"),
-    xfail("special.spherical_bessel_j0"),
-    xfail("special.xlog1py"),
-    xfail("special.zeta"),
-    xfail("squeeze_copy"),
-    xfail("std_mean"),
-    xfail("std_mean", "unbiased"),
-    xfail("stft"),
-    xfail("take"),
-    xfail("to_sparse"),
-    xfail("triangular_solve"),
-    xfail("unfold"),
-    xfail("unfold_copy"),
-    xfail("unique"),
-    xfail("unique_consecutive"),
-    xfail("unsafe_split"),
-    xfail("var_mean"),
-    xfail("var_mean", "unbiased"),
-}
 
 # Add a list of ops that are currently failing BW pass
 skip_bw = [
@@ -827,7 +385,7 @@ class TestMultiThreadedDTensorOps(DTensorOpTestBase, TestDTensorOps):
         _op_db,
         "TestMultiThreadedDTensorOps",
         "test_dtensor_op_db",
-        dtensor_fails | dtensor_multi_threaded_fails | dtensor_fails_no_strategy,
+        dtensor_xfails | no_strategy_xfails | multi_threaded_xfails,
     )
     def test_dtensor_op_db(self, dtype, op):
         self.run_opinfo_test(dtype, op)
@@ -865,7 +423,7 @@ class TestLocalDTensorOps(TestDTensorOps):
         _op_db,
         "TestLocalDTensorOps",
         "test_dtensor_op_db",
-        dtensor_fails | dtensor_fails_no_strategy,
+        dtensor_xfails | no_strategy_xfails,
     )
     def test_dtensor_op_db(self, dtype, op):
         self.run_opinfo_test(dtype, op)
@@ -883,184 +441,6 @@ class TestLocalDTensorOps(TestDTensorOps):
 
     def assertEqualOnRank(self, x, y, msg=None, *, rank=0):
         self.assertEqual(x, y, msg)
-
-
-# Ops where DTensor shard prop has DDEs with unbacked (base tensor passes).
-# This list only contains ops NOT in ops_dde_xfail - those are base tensor issues.
-ops_unbacked_dtensor_dde = {
-    xfail("__getitem__"),
-    xfail("__radd__"),
-    xfail("__rdiv__"),
-    xfail("__rmatmul__"),
-    xfail("__rmod__"),
-    xfail("__rmul__"),
-    xfail("__rpow__"),
-    xfail("__rsub__"),
-    xfail("_segment_reduce", "lengths"),
-    xfail("_segment_reduce", "offsets"),
-    xfail("_softmax_backward_data"),
-    xfail("_unsafe_masked_index"),
-    xfail("add"),
-    xfail("addcdiv"),
-    xfail("addcmul"),
-    xfail("addmm"),
-    xfail("addmm", "decomposed"),
-    xfail("addr"),
-    xfail("alias_copy"),
-    xfail("aminmax"),
-    xfail("argmax"),
-    xfail("argmin"),
-    xfail("argsort"),
-    xfail("argwhere"),
-    xfail("as_strided"),
-    xfail("as_strided", "partial_views"),
-    xfail("atan2"),
-    xfail("block_diag"),
-    xfail("bmm"),
-    xfail("broadcast_tensors"),
-    skip("broadcast_to"),
-    xfail("bucketize"),
-    xfail("cartesian_prod"),
-    xfail("cat"),
-    xfail("cholesky_solve"),
-    xfail("clamp"),
-    xfail("clamp_max"),
-    xfail("clamp_min"),
-    xfail("column_stack"),
-    xfail("constant_pad_nd"),
-    xfail("copysign"),
-    xfail("cumprod"),
-    xfail("cumsum"),
-    xfail("diag"),
-    xfail("diagflat"),
-    xfail("div", "floor_rounding"),
-    xfail("div", "no_rounding_mode"),
-    xfail("div", "trunc_rounding"),
-    xfail("dot"),
-    xfail("dstack"),
-    xfail("einsum"),
-    xfail("eq"),
-    xfail("expand"),
-    xfail("expand_as"),
-    xfail("expand_copy"),
-    xfail("fill"),
-    xfail("flatten"),
-    xfail("float_power"),
-    xfail("floor_divide"),
-    xfail("fmax"),
-    xfail("fmin"),
-    xfail("fmod"),
-    xfail("gather"),
-    xfail("ge"),
-    xfail("gt"),
-    xfail("heaviside"),
-    xfail("histc"),
-    xfail("hstack"),
-    xfail("hypot"),
-    xfail("igamma"),
-    xfail("igammac"),
-    xfail("index_put"),
-    xfail("index_select"),
-    xfail("isclose"),
-    xfail("ldexp"),
-    xfail("le"),
-    xfail("lerp"),
-    xfail("logaddexp"),
-    xfail("logaddexp2"),
-    xfail("logical_and"),
-    xfail("logical_or"),
-    xfail("logical_xor"),
-    xfail("lt"),
-    xfail("masked.normalize"),
-    xfail("masked_fill"),
-    xfail("masked_scatter"),
-    xfail("masked_select"),
-    xfail("matmul"),
-    xfail("max", "binary"),
-    xfail("max", "reduction_with_dim"),
-    xfail("maximum"),
-    xfail("meshgrid", "list_of_tensors"),
-    xfail("meshgrid", "variadic_tensors"),
-    xfail("min", "binary"),
-    xfail("min", "reduction_with_dim"),
-    xfail("minimum"),
-    xfail("mm"),
-    xfail("msort"),
-    xfail("mul"),
-    xfail("mv"),
-    xfail("narrow"),
-    xfail("native_dropout_backward"),
-    xfail("new_empty"),
-    xfail("new_empty_strided"),
-    xfail("new_full"),
-    xfail("new_ones"),
-    xfail("new_zeros"),
-    xfail("nextafter"),
-    xfail("nn.functional.celu"),
-    xfail("nn.functional.cosine_embedding_loss"),
-    xfail("nn.functional.elu"),
-    xfail("nn.functional.hardsigmoid"),
-    xfail("nn.functional.hardtanh"),
-    xfail("nn.functional.hinge_embedding_loss"),
-    xfail("nn.functional.interpolate", "nearest"),
-    xfail("nn.functional.interpolate", "nearest-exact"),
-    xfail("nn.functional.linear"),
-    xfail("nn.functional.logsigmoid"),
-    xfail("nn.functional.margin_ranking_loss"),
-    xfail("nn.functional.mish"),
-    xfail("nn.functional.multilabel_soft_margin_loss"),
-    xfail("nn.functional.normalize"),
-    xfail("nn.functional.pad", "constant"),
-    xfail("nn.functional.pixel_unshuffle"),
-    xfail("nn.functional.poisson_nll_loss"),
-    xfail("nn.functional.relu6"),
-    xfail("nn.functional.selu"),
-    xfail("nn.functional.softplus"),
-    xfail("nn.functional.soft_margin_loss"),
-    xfail("nn.functional.triplet_margin_loss"),
-    xfail("nn.functional.triplet_margin_with_distance_loss"),
-    xfail("nonzero_static"),
-    xfail("outer"),
-    xfail("permute_copy"),
-    xfail("pow"),
-    xfail("prod"),
-    xfail("ravel"),
-    xfail("remainder"),
-    xfail("reshape"),
-    xfail("reshape_as"),
-    xfail("rsub"),
-    xfail("select"),
-    xfail("slice"),
-    xfail("slice_scatter"),
-    xfail("sort"),
-    xfail("split", "list_args"),
-    xfail("split_with_sizes"),
-    xfail("split_with_sizes_copy"),
-    xfail("squeeze"),
-    xfail("squeeze", "multiple"),
-    xfail("stack"),
-    xfail("std"),
-    xfail("std", "unbiased"),
-    xfail("sub"),
-    xfail("topk"),
-    xfail("torch.ops.aten._safe_softmax.default"),
-    xfail("trace"),
-    xfail("transpose_copy"),
-    xfail("tril"),
-    xfail("triu"),
-    xfail("true_divide"),
-    xfail("unflatten"),
-    xfail("unsqueeze_copy"),
-    xfail("var"),
-    xfail("var", "unbiased"),
-    xfail("vdot"),
-    xfail("view"),
-    xfail("view_as"),
-    xfail("view_copy"),
-    xfail("vstack"),
-    xfail("where"),
-    xfail("xlogy"),
-}
 
 
 class TestUnbackedDTensorOps(TestDTensorOps):
@@ -1170,10 +550,7 @@ class TestUnbackedDTensorOps(TestDTensorOps):
         _op_db,
         "TestUnbackedDTensorOps",
         "test_unbacked_dtensor_op_db",
-        ops_dde_xfail
-        | ops_unbacked_dtensor_dde
-        | dtensor_fails_no_strategy
-        | ops_unbacked_skip,
+        unbacked_dtensor_xfails,
     )
     def test_unbacked_dtensor_op_db(self, dtype, op):
         # Filter to samples with valid unbacked dimensions
@@ -1348,13 +725,7 @@ class TestCompiledDTensorOps(TestDTensorOps):
         _op_db,
         "TestCompiledDTensorOps",
         "test_compiled_dtensor_op_db",
-        (
-            dtensor_fails
-            | dtensor_fails_no_strategy
-            | dtensor_multi_threaded_fails
-            | dtensor_compiled_fails
-        )
-        - dtensor_numeric_only_fails,
+        dtensor_compiled_xfails,
     )
     def test_compiled_dtensor_op_db(self, dtype, op):
         self.run_opinfo_test(dtype, op, requires_grad=False)
