+++ torch/testing/_internal/distributed/_tensor/dtensor_xfails.py
@@ -1,0 +1,1543 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates
+# Owner(s): ["oncall: distributed"]
+
+"""
+Xfail and skip lists for DTensor OpInfo tests.
+
+Structure:
+- no_strategy_xfails: Ops with no sharding strategy (NotImplementedError)
+- dtensor_xfails: Misc DTensor failures (numeric, input generation, etc.)
+- multi_threaded_xfails: Additional failures for multi-threaded execution
+
+- dtensor_compiled_xfails: Complete list for TestCompiledDTensorOps
+- unbacked_dtensor_xfails: DTensor-specific unbacked DDEs
+- unbacked_pass_compiled_fail: Ops that pass unbacked but fail compiled (excluded from union)
+
+- unbacked_base_xfails: Base tensor unbacked (for test_ops_unbacked.py)
+
+Test suite usage:
+- TestLocalDTensorOps: dtensor_xfails | no_strategy_xfails
+- TestMultiThreadedDTensorOps: dtensor_xfails | no_strategy_xfails | multi_threaded_xfails
+- TestCompiledDTensorOps: dtensor_compiled_xfails
+- TestUnbackedDTensorOps: (dtensor_compiled_xfails | unbacked_dtensor_xfails) - unbacked_pass_compiled_fail
+- TestOpsUnbacked: unbacked_base_xfails
+"""
+
+
+def xfail(op_name, variant_name="", *, device_type=None, dtypes=None):
+    return (op_name, variant_name, device_type, dtypes, True)
+
+
+def skip(op_name, variant_name="", *, device_type=None, dtypes=None):
+    return (op_name, variant_name, device_type, dtypes, False)
+
+
+# =============================================================================
+# Ops with no sharding strategy (NotImplementedError during propagation)
+# =============================================================================
+no_strategy_xfails = {
+    xfail("_batch_norm_with_update"),
+    xfail("_native_batch_norm_legit"),
+    xfail("_upsample_bilinear2d_aa"),
+    xfail("addbmm"),
+    xfail("allclose"),
+    xfail("as_strided_copy"),
+    xfail("as_strided_scatter"),
+    xfail("cauchy"),
+    xfail("cdist"),
+    xfail("cholesky"),
+    xfail("cholesky_inverse"),
+    xfail("combinations"),
+    xfail("complex"),
+    xfail("cross"),
+    xfail("cummax"),
+    xfail("cummin"),
+    xfail("diagonal_scatter"),
+    xfail("exponential"),
+    xfail("fft.fft"),
+    xfail("fft.fft2"),
+    xfail("fft.fftn"),
+    xfail("fft.fftshift"),
+    xfail("fft.ifft"),
+    xfail("fft.ifft2"),
+    xfail("fft.ifftshift"),
+    xfail("fft.ihfft"),
+    xfail("fft.ihfft2"),
+    xfail("fft.ihfftn"),
+    xfail("fft.irfft2"),
+    xfail("fft.irfftn"),
+    xfail("fft.rfft"),
+    xfail("fft.rfft2"),
+    xfail("fft.rfftn"),
+    xfail("flip"),
+    xfail("fliplr"),
+    xfail("flipud"),
+    xfail("frexp"),
+    xfail("geometric"),
+    xfail("geqrf"),
+    xfail("grid_sampler_2d"),
+    xfail("histogram"),
+    xfail("histogramdd"),
+    xfail("index_add"),
+    xfail("index_copy"),
+    xfail("index_fill"),
+    xfail("index_reduce", "amax"),
+    xfail("index_reduce", "amin"),
+    xfail("index_reduce", "mean"),
+    xfail("index_reduce", "prod"),
+    xfail("isin"),
+    xfail("kthvalue"),
+    xfail("linalg.cholesky"),
+    xfail("linalg.cholesky_ex"),
+    xfail("linalg.cross"),
+    xfail("linalg.det"),
+    xfail("linalg.eig"),
+    xfail("linalg.eigvals"),
+    xfail("linalg.householder_product"),
+    xfail("linalg.inv"),
+    xfail("linalg.inv_ex"),
+    xfail("linalg.ldl_factor"),
+    xfail("linalg.ldl_factor_ex"),
+    xfail("linalg.ldl_solve"),
+    xfail("linalg.lstsq"),
+    xfail("linalg.lstsq", "grad_oriented"),
+    xfail("linalg.lu"),
+    xfail("linalg.lu_factor"),
+    xfail("linalg.lu_factor_ex"),
+    xfail("linalg.lu_solve"),
+    xfail("linalg.pinv"),
+    xfail("linalg.pinv", "hermitian"),
+    xfail("linalg.slogdet"),
+    xfail("linalg.solve"),
+    xfail("linalg.solve_ex"),
+    xfail("linalg.solve_triangular"),
+    xfail("linalg.tensorinv"),
+    xfail("linalg.tensorsolve"),
+    xfail("linalg.vander"),
+    xfail("log_normal"),
+    xfail("logcumsumexp"),
+    xfail("logsumexp"),
+    xfail("lu"),
+    xfail("lu_solve"),
+    xfail("lu_unpack"),
+    xfail("matrix_exp"),
+    xfail("max_pool2d_with_indices_backward"),
+    xfail("median"),
+    xfail("mode"),
+    xfail("multinomial"),
+    xfail("nanmean"),
+    xfail("nanmedian"),
+    xfail("nansum"),
+    xfail("narrow_copy"),
+    xfail("native_batch_norm"),
+    xfail("ne"),
+    xfail("nn.functional.adaptive_avg_pool2d"),
+    xfail("nn.functional.adaptive_max_pool1d"),
+    xfail("nn.functional.adaptive_max_pool2d"),
+    xfail("nn.functional.adaptive_max_pool3d"),
+    xfail("nn.functional.avg_pool1d"),
+    xfail("nn.functional.avg_pool2d"),
+    xfail("nn.functional.avg_pool3d"),
+    xfail("nn.functional.batch_norm"),
+    xfail("nn.functional.batch_norm", "without_cudnn"),
+    xfail("nn.functional.bilinear"),
+    xfail("nn.functional.cosine_similarity"),
+    xfail("nn.functional.ctc_loss"),
+    xfail("nn.functional.glu"),
+    xfail("nn.functional.grid_sample"),
+    xfail("nn.functional.group_norm"),
+    xfail("nn.functional.hardshrink"),
+    xfail("nn.functional.instance_norm"),
+    xfail("nn.functional.local_response_norm"),
+    xfail("nn.functional.max_pool1d"),
+    xfail("nn.functional.max_pool2d"),
+    xfail("nn.functional.max_pool3d"),
+    xfail("nn.functional.max_unpool1d"),
+    xfail("nn.functional.max_unpool1d", "grad"),
+    xfail("nn.functional.max_unpool2d"),
+    xfail("nn.functional.max_unpool2d", "grad"),
+    xfail("nn.functional.max_unpool3d"),
+    xfail("nn.functional.max_unpool3d", "grad"),
+    xfail("nn.functional.multi_margin_loss"),
+    xfail("nn.functional.multilabel_margin_loss"),
+    xfail("nn.functional.pdist"),
+    xfail("nn.functional.rrelu"),
+    xfail("nn.functional.threshold"),
+    xfail("nn.functional.upsample_nearest"),
+    xfail("nonzero"),
+    xfail("ormqr"),
+    xfail("pinverse"),
+    xfail("polar"),
+    xfail("put"),
+    xfail("renorm"),
+    xfail("resize_"),
+    xfail("resize_as_"),
+    xfail("rot90"),
+    xfail("scatter_reduce", "amax"),
+    xfail("scatter_reduce", "amin"),
+    xfail("scatter_reduce", "mean"),
+    xfail("scatter_reduce", "prod"),
+    xfail("scatter_reduce", "sum"),
+    xfail("select_scatter"),
+    xfail("special.airy_ai"),
+    xfail("special.bessel_j0"),
+    xfail("special.bessel_j1"),
+    xfail("special.bessel_y0"),
+    xfail("special.bessel_y1"),
+    xfail("special.chebyshev_polynomial_t"),
+    xfail("special.chebyshev_polynomial_u"),
+    xfail("special.chebyshev_polynomial_v"),
+    xfail("special.chebyshev_polynomial_w"),
+    xfail("special.entr"),
+    xfail("special.erfcx"),
+    xfail("special.hermite_polynomial_h"),
+    xfail("special.hermite_polynomial_he"),
+    xfail("special.i0e"),
+    xfail("special.i1"),
+    xfail("special.i1e"),
+    xfail("special.laguerre_polynomial_l"),
+    xfail("special.legendre_polynomial_p"),
+    xfail("special.log_ndtr"),
+    xfail("special.modified_bessel_i0"),
+    xfail("special.modified_bessel_i1"),
+    xfail("special.modified_bessel_k0"),
+    xfail("special.modified_bessel_k1"),
+    xfail("special.ndtri"),
+    xfail("special.scaled_modified_bessel_k0"),
+    xfail("special.scaled_modified_bessel_k1"),
+    xfail("special.shifted_chebyshev_polynomial_t"),
+    xfail("special.shifted_chebyshev_polynomial_u"),
+    xfail("special.shifted_chebyshev_polynomial_v"),
+    xfail("special.shifted_chebyshev_polynomial_w"),
+    xfail("special.spherical_bessel_j0"),
+    xfail("special.xlog1py"),
+    xfail("special.zeta"),
+    xfail("squeeze_copy"),
+    xfail("std_mean"),
+    xfail("std_mean", "unbiased"),
+    xfail("stft"),
+    xfail("take"),
+    xfail("to_sparse"),
+    xfail("triangular_solve"),
+    xfail("unfold"),
+    xfail("unfold_copy"),
+    xfail("unique"),
+    xfail("unique_consecutive"),
+    xfail("unsafe_split"),
+    xfail("var_mean"),
+    xfail("var_mean", "unbiased"),
+}
+
+
+# =============================================================================
+# Misc DTensor failures (numeric, input generation, sharding issues)
+# =============================================================================
+dtensor_xfails = {
+    xfail("__getitem__"),
+    xfail("_chunk_cat"),
+    xfail("_unsafe_masked_index"),
+    xfail("_unsafe_masked_index_put_accumulate"),
+    xfail("arange"),
+    xfail("as_strided"),
+    xfail("as_strided", "partial_views"),
+    xfail("bernoulli"),
+    xfail("block_diag"),
+    xfail("broadcast_shapes"),
+    xfail("cartesian_prod"),
+    xfail("cholesky_solve"),
+    xfail("eye"),
+    xfail("flatten"),
+    xfail("full"),
+    xfail("index_put"),
+    xfail("kron"),
+    xfail("linalg.matrix_power"),
+    xfail("linspace"),
+    xfail("linspace", "tensor_overload"),
+    xfail("logdet"),
+    xfail("logspace"),
+    xfail("logspace", "tensor_overload"),
+    xfail("masked.logsumexp"),
+    xfail("masked.median"),
+    xfail("masked_scatter"),
+    xfail("masked_select"),
+    xfail("nanquantile"),
+    xfail("nn.functional.adaptive_avg_pool1d"),
+    xfail("nn.functional.adaptive_avg_pool3d"),
+    xfail("nn.functional.alpha_dropout"),
+    xfail("nn.functional.conv1d"),
+    xfail("nn.functional.conv2d"),
+    xfail("nn.functional.conv3d"),
+    xfail("nn.functional.conv_transpose1d"),
+    xfail("nn.functional.conv_transpose2d"),
+    xfail("nn.functional.conv_transpose3d"),
+    xfail("nn.functional.dropout"),
+    xfail("nn.functional.fractional_max_pool2d"),
+    xfail("nn.functional.fractional_max_pool3d"),
+    xfail("nn.functional.huber_loss"),
+    xfail("nn.functional.interpolate", "area"),
+    xfail("nn.functional.interpolate", "nearest"),
+    xfail("nn.functional.interpolate", "nearest-exact"),
+    xfail("nn.functional.pad", "reflect"),
+    xfail("nn.functional.pad", "replicate"),
+    xfail("nn.functional.pad", "replicate_negative"),
+    xfail("nn.functional.unfold"),
+    xfail("normal"),
+    xfail("normal", "in_place"),
+    xfail("normal", "number_mean"),
+    xfail("ones"),
+    xfail("pca_lowrank"),
+    xfail("quantile"),
+    xfail("rand_like"),
+    xfail("randint"),
+    xfail("randint_like"),
+    xfail("randn"),
+    xfail("randn_like"),
+    xfail("ravel"),
+    xfail("repeat_interleave"),
+    xfail("reshape"),
+    xfail("reshape_as"),
+    xfail("roll"),
+    xfail("scalar_tensor"),
+    xfail("searchsorted"),
+    xfail("signal.windows.bartlett"),
+    xfail("signal.windows.blackman"),
+    xfail("signal.windows.cosine"),
+    xfail("signal.windows.exponential"),
+    xfail("signal.windows.gaussian"),
+    xfail("signal.windows.general_cosine"),
+    xfail("signal.windows.general_hamming"),
+    xfail("signal.windows.hamming"),
+    xfail("signal.windows.hann"),
+    xfail("signal.windows.kaiser"),
+    xfail("signal.windows.nuttall"),
+    xfail("sparse.mm", "reduce"),
+    xfail("sparse.sampled_addmm"),
+    xfail("squeeze", "multiple"),
+    xfail("stack"),
+    xfail("svd_lowrank"),
+    xfail("take_along_dim"),
+    xfail("tensor_split"),
+    xfail("transpose"),
+    xfail("unbind"),
+    xfail("unflatten"),
+    xfail("uniform"),
+    xfail("unsafe_chunk"),
+    xfail("view"),
+    xfail("view_as"),
+    xfail("zeros"),
+    # skips
+    skip("__rmatmul__"),
+    skip("argwhere"),
+    skip("cumprod"),
+    skip("empty"),
+    skip("empty_like"),
+    skip("empty_permuted"),
+    skip("empty_strided"),
+    skip("fft.hfft"),
+    skip("fft.hfft2"),
+    skip("fft.hfftn"),
+    skip("fft.ifftn"),
+    skip("fft.irfft"),
+    skip("isclose"),
+    skip("isreal"),
+    skip("istft"),
+    skip("masked.mean"),
+    skip("masked.normalize"),
+    skip("masked.std"),
+    skip("masked.var"),
+    skip("matmul"),
+    skip("meshgrid", "list_of_tensors"),
+    skip("meshgrid", "variadic_tensors"),
+    skip("new_empty"),
+    skip("new_empty_strided"),
+    skip("nn.functional.cosine_embedding_loss"),
+    skip("nn.functional.embedding"),
+    skip("nn.functional.embedding_bag"),
+    skip("nn.functional.feature_alpha_dropout", "with_train"),
+    skip("nn.functional.feature_alpha_dropout", "without_train"),
+    skip("nn.functional.hinge_embedding_loss"),
+    skip("nn.functional.scaled_dot_product_attention"),
+    skip("nn.functional.softmin"),
+    skip("prod"),
+    skip("_segment_reduce", "lengths"),
+    skip("_segment_reduce", "offsets"),
+    skip("squeeze"),
+}
+
+
+# =============================================================================
+# Additional failures for multi-threaded execution only
+# =============================================================================
+multi_threaded_xfails = {
+    xfail("full_like"),
+    xfail("masked.cumprod"),
+    xfail("nn.functional.dropout2d"),
+    xfail("nn.functional.dropout3d"),
+    skip("nn.functional.multi_head_attention_forward"),
+}
+
+
+# =============================================================================
+# Complete list for TestCompiledDTensorOps
+# =============================================================================
+dtensor_compiled_xfails = {
+    xfail("__getitem__"),
+    xfail("_batch_norm_with_update"),
+    xfail("_chunk_cat"),
+    xfail("_native_batch_norm_legit"),
+    xfail("_unsafe_masked_index"),
+    xfail("_unsafe_masked_index_put_accumulate"),
+    xfail("_upsample_bilinear2d_aa"),
+    xfail("addbmm"),
+    xfail("allclose"),
+    xfail("argmax"),
+    xfail("argmin"),
+    xfail("as_strided"),
+    xfail("as_strided", "partial_views"),
+    xfail("as_strided_copy"),
+    xfail("as_strided_scatter"),
+    xfail("atleast_1d"),
+    xfail("atleast_2d"),
+    xfail("atleast_3d"),
+    xfail("bernoulli"),
+    xfail("block_diag"),
+    xfail("broadcast_tensors"),
+    xfail("broadcast_to"),
+    xfail("cartesian_prod"),
+    xfail("cauchy"),
+    xfail("cdist"),
+    xfail("cholesky"),
+    xfail("cholesky_inverse"),
+    xfail("cholesky_solve"),
+    xfail("combinations"),
+    xfail("complex"),
+    xfail("corrcoef"),
+    xfail("cov"),
+    xfail("cross"),
+    xfail("cummax"),
+    xfail("cummin"),
+    xfail("diagonal"),
+    xfail("diagonal_scatter"),
+    xfail("dsplit"),
+    xfail("equal"),
+    xfail("expand"),
+    xfail("expand_as"),
+    xfail("exponential"),
+    xfail("fft.fft"),
+    xfail("fft.fft2"),
+    xfail("fft.fftn"),
+    xfail("fft.fftshift"),
+    xfail("fft.ifft"),
+    xfail("fft.ifft2"),
+    xfail("fft.ifftshift"),
+    xfail("fft.ihfft"),
+    xfail("fft.ihfft2"),
+    xfail("fft.ihfftn"),
+    xfail("fft.irfft2"),
+    xfail("fft.irfftn"),
+    xfail("fft.rfft"),
+    xfail("fft.rfft2"),
+    xfail("fft.rfftn"),
+    xfail("flatten"),
+    xfail("flip"),
+    xfail("fliplr"),
+    xfail("flipud"),
+    xfail("frexp"),
+    xfail("gather"),
+    xfail("geometric"),
+    xfail("geqrf"),
+    xfail("grid_sampler_2d"),
+    xfail("hash_tensor"),
+    xfail("histogram"),
+    xfail("histogramdd"),
+    xfail("hsplit"),
+    xfail("index_add"),
+    xfail("index_copy"),
+    xfail("index_fill"),
+    xfail("index_put"),
+    xfail("index_reduce", "amax"),
+    xfail("index_reduce", "amin"),
+    xfail("index_reduce", "mean"),
+    xfail("index_reduce", "prod"),
+    xfail("index_select"),
+    xfail("isin"),
+    xfail("item"),
+    xfail("kron"),
+    xfail("kthvalue"),
+    xfail("linalg.cholesky"),
+    xfail("linalg.cholesky_ex"),
+    xfail("linalg.cross"),
+    xfail("linalg.det"),
+    xfail("linalg.diagonal"),
+    xfail("linalg.eig"),
+    xfail("linalg.eigvals"),
+    xfail("linalg.householder_product"),
+    xfail("linalg.inv"),
+    xfail("linalg.inv_ex"),
+    xfail("linalg.ldl_factor"),
+    xfail("linalg.ldl_factor_ex"),
+    xfail("linalg.ldl_solve"),
+    xfail("linalg.lstsq"),
+    xfail("linalg.lstsq", "grad_oriented"),
+    xfail("linalg.lu"),
+    xfail("linalg.lu_factor"),
+    xfail("linalg.lu_factor_ex"),
+    xfail("linalg.lu_solve"),
+    xfail("linalg.matrix_power"),
+    xfail("linalg.pinv"),
+    xfail("linalg.pinv", "hermitian"),
+    xfail("linalg.slogdet"),
+    xfail("linalg.solve"),
+    xfail("linalg.solve_ex"),
+    xfail("linalg.solve_triangular"),
+    xfail("linalg.tensorinv"),
+    xfail("linalg.tensorsolve"),
+    xfail("linalg.vander"),
+    xfail("linspace", "tensor_overload"),
+    xfail("log_normal"),
+    xfail("logcumsumexp"),
+    xfail("logdet"),
+    xfail("logspace", "tensor_overload"),
+    xfail("logsumexp"),
+    xfail("lu"),
+    xfail("lu_solve"),
+    xfail("lu_unpack"),
+    xfail("masked.argmax"),
+    xfail("masked.argmin"),
+    xfail("masked.cumprod"),
+    xfail("masked.logsumexp"),
+    xfail("masked.median"),
+    xfail("masked_scatter"),
+    xfail("masked_select"),
+    xfail("matrix_exp"),
+    xfail("max", "reduction_with_dim"),
+    xfail("max_pool2d_with_indices_backward"),
+    xfail("median"),
+    xfail("min", "reduction_with_dim"),
+    xfail("mode"),
+    xfail("movedim"),
+    xfail("multinomial"),
+    xfail("nanmean"),
+    xfail("nanmedian"),
+    xfail("nanquantile"),
+    xfail("nansum"),
+    xfail("narrow"),
+    xfail("native_batch_norm"),
+    xfail("ne"),
+    xfail("nn.functional.adaptive_avg_pool1d"),
+    xfail("nn.functional.adaptive_avg_pool2d"),
+    xfail("nn.functional.adaptive_avg_pool3d"),
+    xfail("nn.functional.adaptive_max_pool1d"),
+    xfail("nn.functional.adaptive_max_pool2d"),
+    xfail("nn.functional.adaptive_max_pool3d"),
+    xfail("nn.functional.alpha_dropout"),
+    xfail("nn.functional.avg_pool1d"),
+    xfail("nn.functional.avg_pool2d"),
+    xfail("nn.functional.avg_pool3d"),
+    xfail("nn.functional.batch_norm"),
+    xfail("nn.functional.batch_norm", "without_cudnn"),
+    xfail("nn.functional.bilinear"),
+    xfail("nn.functional.binary_cross_entropy"),
+    xfail("nn.functional.binary_cross_entropy_with_logits"),
+    xfail("nn.functional.conv1d"),
+    xfail("nn.functional.conv2d"),
+    xfail("nn.functional.conv3d"),
+    xfail("nn.functional.conv_transpose1d"),
+    xfail("nn.functional.conv_transpose2d"),
+    xfail("nn.functional.conv_transpose3d"),
+    xfail("nn.functional.cosine_similarity"),
+    xfail("nn.functional.ctc_loss"),
+    xfail("nn.functional.dropout"),
+    xfail("nn.functional.dropout2d"),
+    xfail("nn.functional.dropout3d"),
+    xfail("nn.functional.fractional_max_pool2d"),
+    xfail("nn.functional.fractional_max_pool3d"),
+    xfail("nn.functional.gaussian_nll_loss"),
+    xfail("nn.functional.grid_sample"),
+    xfail("nn.functional.group_norm"),
+    xfail("nn.functional.instance_norm"),
+    xfail("nn.functional.interpolate", "area"),
+    xfail("nn.functional.interpolate", "bicubic"),
+    xfail("nn.functional.interpolate", "bilinear"),
+    xfail("nn.functional.interpolate", "linear"),
+    xfail("nn.functional.interpolate", "nearest"),
+    xfail("nn.functional.interpolate", "nearest-exact"),
+    xfail("nn.functional.interpolate", "trilinear"),
+    xfail("nn.functional.local_response_norm"),
+    xfail("nn.functional.logsigmoid"),
+    xfail("nn.functional.margin_ranking_loss"),
+    xfail("nn.functional.max_pool1d"),
+    xfail("nn.functional.max_pool2d"),
+    xfail("nn.functional.max_pool3d"),
+    xfail("nn.functional.max_unpool1d"),
+    xfail("nn.functional.max_unpool1d", "grad"),
+    xfail("nn.functional.max_unpool2d"),
+    xfail("nn.functional.max_unpool2d", "grad"),
+    xfail("nn.functional.max_unpool3d"),
+    xfail("nn.functional.max_unpool3d", "grad"),
+    xfail("nn.functional.multi_margin_loss"),
+    xfail("nn.functional.multilabel_margin_loss"),
+    xfail("nn.functional.multilabel_soft_margin_loss"),
+    xfail("nn.functional.pad", "reflect"),
+    xfail("nn.functional.pad", "replicate"),
+    xfail("nn.functional.pad", "replicate_negative"),
+    xfail("nn.functional.pdist"),
+    xfail("nn.functional.rrelu"),
+    xfail("nn.functional.threshold"),
+    xfail("nn.functional.unfold"),
+    xfail("nn.functional.upsample_bilinear"),
+    xfail("nn.functional.upsample_nearest"),
+    xfail("nonzero"),
+    xfail("nonzero_static"),
+    xfail("normal"),
+    xfail("normal", "in_place"),
+    xfail("normal", "number_mean"),
+    xfail("ormqr"),
+    xfail("pca_lowrank"),
+    xfail("permute"),
+    xfail("pinverse"),
+    xfail("polar"),
+    xfail("put"),
+    xfail("quantile"),
+    xfail("rand_like"),
+    xfail("randint_like"),
+    xfail("randn_like"),
+    xfail("ravel"),
+    xfail("renorm"),
+    xfail("repeat_interleave"),
+    xfail("reshape"),
+    xfail("reshape_as"),
+    xfail("resize_"),
+    xfail("resize_as_"),
+    xfail("roll"),
+    xfail("rot90"),
+    xfail("scatter"),
+    xfail("scatter_add"),
+    xfail("scatter_reduce", "amax"),
+    xfail("scatter_reduce", "amin"),
+    xfail("scatter_reduce", "mean"),
+    xfail("scatter_reduce", "prod"),
+    xfail("scatter_reduce", "sum"),
+    xfail("searchsorted"),
+    xfail("select"),
+    xfail("select_scatter"),
+    xfail("slice"),
+    xfail("sparse.sampled_addmm"),
+    xfail("special.airy_ai"),
+    xfail("special.bessel_j0"),
+    xfail("special.bessel_j1"),
+    xfail("special.bessel_y0"),
+    xfail("special.bessel_y1"),
+    xfail("special.chebyshev_polynomial_t"),
+    xfail("special.chebyshev_polynomial_u"),
+    xfail("special.chebyshev_polynomial_v"),
+    xfail("special.chebyshev_polynomial_w"),
+    xfail("special.entr"),
+    xfail("special.erfcx"),
+    xfail("special.hermite_polynomial_h"),
+    xfail("special.hermite_polynomial_he"),
+    xfail("special.i0e"),
+    xfail("special.i1"),
+    xfail("special.i1e"),
+    xfail("special.laguerre_polynomial_l"),
+    xfail("special.legendre_polynomial_p"),
+    xfail("special.log_ndtr"),
+    xfail("special.modified_bessel_i0"),
+    xfail("special.modified_bessel_i1"),
+    xfail("special.modified_bessel_k0"),
+    xfail("special.modified_bessel_k1"),
+    xfail("special.ndtri"),
+    xfail("special.scaled_modified_bessel_k0"),
+    xfail("special.scaled_modified_bessel_k1"),
+    xfail("special.shifted_chebyshev_polynomial_t"),
+    xfail("special.shifted_chebyshev_polynomial_u"),
+    xfail("special.shifted_chebyshev_polynomial_v"),
+    xfail("special.shifted_chebyshev_polynomial_w"),
+    xfail("special.spherical_bessel_j0"),
+    xfail("special.xlog1py"),
+    xfail("special.zeta"),
+    xfail("squeeze", "multiple"),
+    xfail("std_mean"),
+    xfail("std_mean", "unbiased"),
+    xfail("stft"),
+    xfail("svd_lowrank"),
+    xfail("t"),
+    xfail("take"),
+    xfail("take_along_dim"),
+    xfail("tensor_split"),
+    xfail("to_sparse"),
+    xfail("transpose"),
+    xfail("transpose_copy"),
+    xfail("triangular_solve"),
+    xfail("unbind"),
+    xfail("unflatten"),
+    xfail("unfold"),
+    xfail("unfold_copy"),
+    xfail("uniform"),
+    xfail("unique"),
+    xfail("unique_consecutive"),
+    xfail("unsafe_split"),
+    xfail("unsqueeze"),
+    xfail("var_mean"),
+    xfail("var_mean", "unbiased"),
+    xfail("view"),
+    xfail("view_as"),
+    xfail("vsplit"),
+    # skips
+    skip("__rmatmul__"),
+    skip("argwhere"),
+    skip("cumprod"),
+    skip("empty"),
+    skip("empty_like"),
+    skip("empty_permuted"),
+    skip("empty_strided"),
+    skip("fft.hfft"),
+    skip("fft.hfft2"),
+    skip("fft.hfftn"),
+    skip("fft.ifftn"),
+    skip("fft.irfft"),
+    skip("isclose"),
+    skip("isreal"),
+    skip("istft"),
+    skip("masked.mean"),
+    skip("masked.normalize"),
+    skip("masked.std"),
+    skip("masked.var"),
+    skip("matmul"),
+    skip("meshgrid", "list_of_tensors"),
+    skip("meshgrid", "variadic_tensors"),
+    skip("new_empty"),
+    skip("new_empty_strided"),
+    skip("nn.functional.cosine_embedding_loss"),
+    skip("nn.functional.embedding"),
+    skip("nn.functional.embedding_bag"),
+    skip("nn.functional.feature_alpha_dropout", "with_train"),
+    skip("nn.functional.feature_alpha_dropout", "without_train"),
+    skip("nn.functional.hinge_embedding_loss"),
+    skip("nn.functional.multi_head_attention_forward"),
+    skip("nn.functional.scaled_dot_product_attention"),
+    skip("nn.functional.softmin"),
+    skip("prod"),
+    skip("_segment_reduce", "lengths"),
+    skip("_segment_reduce", "offsets"),
+    skip("squeeze"),
+}
+
+
+# =============================================================================
+# Unbacked DTensor xfails for TestUnbackedDTensorOps
+# =============================================================================
+unbacked_dtensor_xfails = {
+    xfail("__getitem__"),
+    xfail("__radd__"),
+    xfail("__rdiv__"),
+    xfail("__rmatmul__"),
+    xfail("__rmod__"),
+    xfail("__rmul__"),
+    xfail("__rpow__"),
+    xfail("__rsub__"),
+    xfail("_batch_norm_with_update"),
+    xfail("_chunk_cat"),
+    xfail("_native_batch_norm_legit"),
+    xfail("_segment_reduce", "lengths"),
+    xfail("_segment_reduce", "offsets"),
+    xfail("_softmax_backward_data"),
+    xfail("_unsafe_masked_index"),
+    xfail("_unsafe_masked_index_put_accumulate"),
+    xfail("_upsample_bilinear2d_aa"),
+    xfail("add"),
+    xfail("addbmm"),
+    xfail("addcdiv"),
+    xfail("addcmul"),
+    xfail("addmm"),
+    xfail("addmm", "decomposed"),
+    xfail("addmv"),
+    xfail("addr"),
+    xfail("alias_copy"),
+    xfail("allclose"),
+    xfail("aminmax"),
+    xfail("argmax"),
+    xfail("argmin"),
+    xfail("argsort"),
+    xfail("argwhere"),
+    xfail("as_strided"),
+    xfail("as_strided", "partial_views"),
+    xfail("as_strided_copy"),
+    xfail("as_strided_scatter"),
+    xfail("atan2"),
+    xfail("baddbmm"),
+    xfail("bernoulli"),
+    xfail("block_diag"),
+    xfail("bmm"),
+    xfail("broadcast_tensors"),
+    xfail("broadcast_to"),
+    xfail("bucketize"),
+    xfail("cartesian_prod"),
+    xfail("cat"),
+    xfail("cauchy"),
+    xfail("cdist"),
+    xfail("cholesky"),
+    xfail("cholesky_inverse"),
+    xfail("cholesky_solve"),
+    xfail("chunk"),
+    xfail("clamp"),
+    xfail("clamp_max"),
+    xfail("clamp_min"),
+    xfail("column_stack"),
+    xfail("combinations"),
+    xfail("complex"),
+    xfail("constant_pad_nd"),
+    xfail("copysign"),
+    xfail("corrcoef"),
+    xfail("cov"),
+    xfail("cross"),
+    xfail("cummax"),
+    xfail("cummin"),
+    xfail("cumprod"),
+    xfail("cumsum"),
+    xfail("cumulative_trapezoid"),
+    xfail("diag"),
+    xfail("diagflat"),
+    xfail("diagonal_scatter"),
+    xfail("diff"),
+    xfail("dist"),
+    xfail("div", "floor_rounding"),
+    xfail("div", "no_rounding_mode"),
+    xfail("div", "trunc_rounding"),
+    xfail("dot"),
+    xfail("dsplit"),
+    xfail("dstack"),
+    xfail("einsum"),
+    xfail("eq"),
+    xfail("equal"),
+    xfail("expand"),
+    xfail("expand_as"),
+    xfail("expand_copy"),
+    xfail("exponential"),
+    xfail("fft.fft"),
+    xfail("fft.fft2"),
+    xfail("fft.fftn"),
+    xfail("fft.fftshift"),
+    xfail("fft.ifft"),
+    xfail("fft.ifft2"),
+    xfail("fft.ifftshift"),
+    xfail("fft.ihfft"),
+    xfail("fft.ihfft2"),
+    xfail("fft.ihfftn"),
+    xfail("fft.irfft2"),
+    xfail("fft.irfftn"),
+    xfail("fft.rfft"),
+    xfail("fft.rfft2"),
+    xfail("fft.rfftn"),
+    xfail("fill"),
+    xfail("flatten"),
+    xfail("flip"),
+    xfail("fliplr"),
+    xfail("flipud"),
+    xfail("float"),
+    xfail("float_power"),
+    xfail("floor_divide"),
+    xfail("fmax"),
+    xfail("fmin"),
+    xfail("fmod"),
+    xfail("frexp"),
+    xfail("gather"),
+    xfail("ge"),
+    xfail("geometric"),
+    xfail("geqrf"),
+    xfail("gradient"),
+    xfail("grid_sampler_2d"),
+    xfail("gt"),
+    xfail("hash_tensor"),
+    xfail("heaviside"),
+    xfail("histc"),
+    xfail("histogram"),
+    xfail("histogramdd"),
+    xfail("hsplit"),
+    xfail("hstack"),
+    xfail("hypot"),
+    xfail("igamma"),
+    xfail("igammac"),
+    xfail("index_add"),
+    xfail("index_copy"),
+    xfail("index_fill"),
+    xfail("index_put"),
+    xfail("index_reduce", "amax"),
+    xfail("index_reduce", "amin"),
+    xfail("index_reduce", "mean"),
+    xfail("index_reduce", "prod"),
+    xfail("index_select"),
+    xfail("inner"),
+    xfail("isclose"),
+    xfail("isin"),
+    xfail("item"),
+    xfail("kron"),
+    xfail("kthvalue"),
+    xfail("ldexp"),
+    xfail("le"),
+    xfail("lerp"),
+    xfail("linalg.cholesky"),
+    xfail("linalg.cholesky_ex"),
+    xfail("linalg.cond"),
+    xfail("linalg.cross"),
+    xfail("linalg.det"),
+    xfail("linalg.eig"),
+    xfail("linalg.eigvals"),
+    xfail("linalg.householder_product"),
+    xfail("linalg.inv"),
+    xfail("linalg.inv_ex"),
+    xfail("linalg.ldl_factor"),
+    xfail("linalg.ldl_factor_ex"),
+    xfail("linalg.ldl_solve"),
+    xfail("linalg.lstsq"),
+    xfail("linalg.lstsq", "grad_oriented"),
+    xfail("linalg.lu"),
+    xfail("linalg.lu_factor"),
+    xfail("linalg.lu_factor_ex"),
+    xfail("linalg.lu_solve"),
+    xfail("linalg.matrix_power"),
+    xfail("linalg.matrix_norm"),
+    xfail("linalg.matrix_rank"),
+    xfail("linalg.matrix_rank", "hermitian"),
+    xfail("linalg.multi_dot"),
+    xfail("linalg.norm"),
+    xfail("linalg.norm", "subgradients_at_zero"),
+    xfail("linalg.pinv"),
+    xfail("linalg.pinv", "hermitian"),
+    xfail("linalg.qr"),
+    xfail("linalg.slogdet"),
+    xfail("linalg.solve"),
+    xfail("linalg.solve_ex"),
+    xfail("linalg.solve_triangular"),
+    xfail("linalg.svd"),
+    xfail("linalg.svdvals"),
+    xfail("linalg.tensorinv"),
+    xfail("linalg.tensorsolve"),
+    xfail("linalg.vander"),
+    xfail("linalg.vector_norm"),
+    xfail("linspace", "tensor_overload"),
+    xfail("log_normal"),
+    xfail("logaddexp"),
+    xfail("logaddexp2"),
+    xfail("logcumsumexp"),
+    xfail("logdet"),
+    xfail("logical_and"),
+    xfail("logical_or"),
+    xfail("logical_xor"),
+    xfail("logspace", "tensor_overload"),
+    xfail("logsumexp"),
+    xfail("lt"),
+    xfail("lu"),
+    xfail("lu_solve"),
+    xfail("lu_unpack"),
+    xfail("masked.amax"),
+    xfail("masked.amin"),
+    xfail("masked.argmax"),
+    xfail("masked.argmin"),
+    xfail("masked.cumprod"),
+    xfail("masked.cumsum"),
+    xfail("masked.log_softmax"),
+    xfail("masked.logaddexp"),
+    xfail("masked.logsumexp"),
+    xfail("masked.median"),
+    xfail("masked.norm"),
+    xfail("masked.normalize"),
+    xfail("masked.prod"),
+    xfail("masked.softmax"),
+    xfail("masked.softmin"),
+    xfail("masked.sum"),
+    xfail("masked_fill"),
+    xfail("masked_scatter"),
+    xfail("masked_select"),
+    xfail("matmul"),
+    xfail("matrix_exp"),
+    xfail("max", "binary"),
+    xfail("max", "reduction_with_dim"),
+    xfail("max_pool2d_with_indices_backward"),
+    xfail("maximum"),
+    xfail("median"),
+    xfail("meshgrid", "list_of_tensors"),
+    xfail("meshgrid", "variadic_tensors"),
+    xfail("min", "binary"),
+    xfail("min", "reduction_with_dim"),
+    xfail("minimum"),
+    xfail("mm"),
+    xfail("mode"),
+    xfail("msort"),
+    xfail("mul"),
+    xfail("multinomial"),
+    xfail("mv"),
+    xfail("nanmean"),
+    xfail("nanmedian"),
+    xfail("nanquantile"),
+    xfail("nansum"),
+    xfail("narrow"),
+    xfail("narrow_copy"),
+    xfail("native_batch_norm"),
+    xfail("native_dropout_backward"),
+    xfail("ne"),
+    xfail("new_empty"),
+    xfail("new_empty_strided"),
+    xfail("new_full"),
+    xfail("new_ones"),
+    xfail("new_zeros"),
+    xfail("nextafter"),
+    xfail("nn.functional.adaptive_avg_pool1d"),
+    xfail("nn.functional.adaptive_avg_pool2d"),
+    xfail("nn.functional.adaptive_avg_pool3d"),
+    xfail("nn.functional.adaptive_max_pool1d"),
+    xfail("nn.functional.adaptive_max_pool2d"),
+    xfail("nn.functional.adaptive_max_pool3d"),
+    xfail("nn.functional.alpha_dropout"),
+    xfail("nn.functional.avg_pool1d"),
+    xfail("nn.functional.avg_pool2d"),
+    xfail("nn.functional.avg_pool3d"),
+    xfail("nn.functional.batch_norm"),
+    xfail("nn.functional.batch_norm", "without_cudnn"),
+    xfail("nn.functional.bilinear"),
+    xfail("nn.functional.binary_cross_entropy"),
+    xfail("nn.functional.binary_cross_entropy_with_logits"),
+    xfail("nn.functional.celu"),
+    xfail("nn.functional.channel_shuffle"),
+    xfail("nn.functional.conv1d"),
+    xfail("nn.functional.conv2d"),
+    xfail("nn.functional.conv3d"),
+    xfail("nn.functional.conv_transpose1d"),
+    xfail("nn.functional.conv_transpose2d"),
+    xfail("nn.functional.conv_transpose3d"),
+    xfail("nn.functional.cosine_embedding_loss"),
+    xfail("nn.functional.cosine_similarity"),
+    xfail("nn.functional.cross_entropy"),
+    xfail("nn.functional.ctc_loss"),
+    xfail("nn.functional.dropout"),
+    xfail("nn.functional.dropout2d"),
+    xfail("nn.functional.dropout3d"),
+    xfail("nn.functional.elu"),
+    xfail("nn.functional.fractional_max_pool2d"),
+    xfail("nn.functional.fractional_max_pool3d"),
+    xfail("nn.functional.gaussian_nll_loss"),
+    xfail("nn.functional.glu"),
+    xfail("nn.functional.grid_sample"),
+    xfail("nn.functional.group_norm"),
+    xfail("nn.functional.hardsigmoid"),
+    xfail("nn.functional.hardshrink"),
+    xfail("nn.functional.hardtanh"),
+    xfail("nn.functional.hinge_embedding_loss"),
+    xfail("nn.functional.huber_loss"),
+    xfail("nn.functional.instance_norm"),
+    xfail("nn.functional.interpolate", "area"),
+    xfail("nn.functional.interpolate", "bicubic"),
+    xfail("nn.functional.interpolate", "bilinear"),
+    xfail("nn.functional.interpolate", "linear"),
+    xfail("nn.functional.interpolate", "nearest"),
+    xfail("nn.functional.interpolate", "nearest-exact"),
+    xfail("nn.functional.interpolate", "trilinear"),
+    xfail("nn.functional.l1_loss"),
+    xfail("nn.functional.linear"),
+    xfail("nn.functional.local_response_norm"),
+    xfail("nn.functional.logsigmoid"),
+    xfail("nn.functional.margin_ranking_loss"),
+    xfail("nn.functional.max_pool1d"),
+    xfail("nn.functional.max_pool2d"),
+    xfail("nn.functional.max_pool3d"),
+    xfail("nn.functional.max_unpool1d"),
+    xfail("nn.functional.max_unpool1d", "grad"),
+    xfail("nn.functional.max_unpool2d"),
+    xfail("nn.functional.max_unpool2d", "grad"),
+    xfail("nn.functional.max_unpool3d"),
+    xfail("nn.functional.max_unpool3d", "grad"),
+    xfail("nn.functional.mish"),
+    xfail("nn.functional.mse_loss"),
+    xfail("nn.functional.multi_margin_loss"),
+    xfail("nn.functional.multilabel_margin_loss"),
+    xfail("nn.functional.multilabel_soft_margin_loss"),
+    xfail("nn.functional.nll_loss"),
+    xfail("nn.functional.normalize"),
+    xfail("nn.functional.pad", "circular"),
+    xfail("nn.functional.pad", "constant"),
+    xfail("nn.functional.pad", "reflect"),
+    xfail("nn.functional.pad", "replicate"),
+    xfail("nn.functional.pad", "replicate_negative"),
+    xfail("nn.functional.pdist"),
+    xfail("nn.functional.pixel_shuffle"),
+    xfail("nn.functional.pixel_unshuffle"),
+    xfail("nn.functional.poisson_nll_loss"),
+    xfail("nn.functional.prelu"),
+    xfail("nn.functional.relu6"),
+    xfail("nn.functional.rrelu"),
+    xfail("nn.functional.selu"),
+    xfail("nn.functional.smooth_l1_loss"),
+    xfail("nn.functional.soft_margin_loss"),
+    xfail("nn.functional.softplus"),
+    xfail("nn.functional.threshold"),
+    xfail("nn.functional.triplet_margin_loss"),
+    xfail("nn.functional.triplet_margin_with_distance_loss"),
+    xfail("nn.functional.unfold"),
+    xfail("nn.functional.upsample_bilinear"),
+    xfail("nn.functional.upsample_nearest"),
+    xfail("nonzero"),
+    xfail("nonzero_static"),
+    xfail("norm"),
+    xfail("norm", "fro"),
+    xfail("norm", "nuc"),
+    xfail("normal"),
+    xfail("normal", "in_place"),
+    xfail("normal", "number_mean"),
+    xfail("ormqr"),
+    xfail("outer"),
+    xfail("pca_lowrank"),
+    xfail("permute_copy"),
+    xfail("pinverse"),
+    xfail("polar"),
+    xfail("pow"),
+    xfail("prod"),
+    xfail("put"),
+    xfail("quantile"),
+    xfail("qr"),
+    xfail("rand_like"),
+    xfail("randint_like"),
+    xfail("randn_like"),
+    xfail("ravel"),
+    xfail("remainder"),
+    xfail("renorm"),
+    xfail("repeat_interleave"),
+    xfail("reshape"),
+    xfail("reshape_as"),
+    xfail("resize_"),
+    xfail("resize_as_"),
+    xfail("roll"),
+    xfail("rot90"),
+    xfail("rsub"),
+    xfail("scatter"),
+    xfail("scatter_add"),
+    xfail("scatter_reduce", "amax"),
+    xfail("scatter_reduce", "amin"),
+    xfail("scatter_reduce", "mean"),
+    xfail("scatter_reduce", "prod"),
+    xfail("scatter_reduce", "sum"),
+    xfail("searchsorted"),
+    xfail("select"),
+    xfail("select_scatter"),
+    xfail("slice"),
+    xfail("slice_scatter"),
+    xfail("sort"),
+    xfail("sparse.mm", "reduce"),
+    xfail("sparse.sampled_addmm"),
+    xfail("special.airy_ai"),
+    xfail("special.bessel_j0"),
+    xfail("special.bessel_j1"),
+    xfail("special.bessel_y0"),
+    xfail("special.bessel_y1"),
+    xfail("special.chebyshev_polynomial_t"),
+    xfail("special.chebyshev_polynomial_u"),
+    xfail("special.chebyshev_polynomial_v"),
+    xfail("special.chebyshev_polynomial_w"),
+    xfail("special.entr"),
+    xfail("special.erfcx"),
+    xfail("special.hermite_polynomial_h"),
+    xfail("special.hermite_polynomial_he"),
+    xfail("special.i0e"),
+    xfail("special.i1"),
+    xfail("special.i1e"),
+    xfail("special.laguerre_polynomial_l"),
+    xfail("special.legendre_polynomial_p"),
+    xfail("special.log_ndtr"),
+    xfail("special.modified_bessel_i0"),
+    xfail("special.modified_bessel_i1"),
+    xfail("special.modified_bessel_k0"),
+    xfail("special.modified_bessel_k1"),
+    xfail("special.ndtri"),
+    xfail("special.scaled_modified_bessel_k0"),
+    xfail("special.scaled_modified_bessel_k1"),
+    xfail("special.shifted_chebyshev_polynomial_t"),
+    xfail("special.shifted_chebyshev_polynomial_u"),
+    xfail("special.shifted_chebyshev_polynomial_v"),
+    xfail("special.shifted_chebyshev_polynomial_w"),
+    xfail("special.spherical_bessel_j0"),
+    xfail("special.xlog1py"),
+    xfail("special.zeta"),
+    xfail("split"),
+    xfail("split", "list_args"),
+    xfail("split_with_sizes"),
+    xfail("split_with_sizes_copy"),
+    xfail("squeeze"),
+    xfail("squeeze", "multiple"),
+    xfail("squeeze_copy"),
+    xfail("stack"),
+    xfail("std"),
+    xfail("std", "unbiased"),
+    xfail("std_mean"),
+    xfail("std_mean", "unbiased"),
+    xfail("stft"),
+    xfail("sub"),
+    xfail("sum_to_size"),
+    xfail("svd"),
+    xfail("svd_lowrank"),
+    xfail("take"),
+    xfail("take_along_dim"),
+    xfail("tensor_split"),
+    xfail("tensordot"),
+    xfail("to_sparse"),
+    xfail("topk"),
+    xfail("torch.ops.aten._safe_softmax.default"),
+    xfail("trace"),
+    xfail("transpose_copy"),
+    xfail("trapezoid"),
+    xfail("trapz"),
+    xfail("triangular_solve"),
+    xfail("tril"),
+    xfail("triu"),
+    xfail("true_divide"),
+    xfail("unbind"),
+    xfail("unbind_copy"),
+    xfail("unflatten"),
+    xfail("unfold"),
+    xfail("unfold_copy"),
+    xfail("uniform"),
+    xfail("unique"),
+    xfail("unique_consecutive"),
+    xfail("unsafe_chunk"),
+    xfail("unsafe_split"),
+    xfail("unsqueeze_copy"),
+    xfail("var"),
+    xfail("var", "unbiased"),
+    xfail("var_mean"),
+    xfail("var_mean", "unbiased"),
+    xfail("vdot"),
+    xfail("view"),
+    xfail("view_as"),
+    xfail("view_copy"),
+    xfail("view_as_complex"),
+    xfail("vsplit"),
+    xfail("vstack"),
+    xfail("where"),
+    xfail("xlogy"),
+    # skips
+    skip("__rmatmul__"),
+    skip("_segment_reduce", "lengths"),
+    skip("_segment_reduce", "offsets"),
+    skip("arange"),
+    skip("argwhere"),
+    skip("broadcast_shapes"),
+    skip("broadcast_to"),
+    skip("cumprod"),
+    skip("empty"),
+    skip("empty_like"),
+    skip("empty_permuted"),
+    skip("empty_strided"),
+    skip("eye"),
+    skip("fft.hfft"),
+    skip("fft.hfft2"),
+    skip("fft.hfftn"),
+    skip("fft.ifftn"),
+    skip("fft.irfft"),
+    skip("full"),
+    skip("isclose"),
+    skip("isreal"),
+    skip("istft"),
+    skip("item"),
+    skip("linspace"),
+    skip("linspace", "tensor_overload"),
+    skip("logspace"),
+    skip("logspace", "tensor_overload"),
+    skip("masked.mean"),
+    skip("masked.normalize"),
+    skip("masked.std"),
+    skip("masked.var"),
+    skip("matmul"),
+    skip("meshgrid", "list_of_tensors"),
+    skip("meshgrid", "variadic_tensors"),
+    skip("new_empty"),
+    skip("new_empty_strided"),
+    skip("nn.functional.cosine_embedding_loss"),
+    skip("nn.functional.embedding"),
+    skip("nn.functional.embedding_bag"),
+    skip("nn.functional.feature_alpha_dropout", "with_train"),
+    skip("nn.functional.feature_alpha_dropout", "without_train"),
+    skip("nn.functional.hinge_embedding_loss"),
+    skip("nn.functional.multi_head_attention_forward"),
+    skip("nn.functional.scaled_dot_product_attention"),
+    skip("nn.functional.softmin"),
+    skip("ones"),
+    skip("prod"),
+    skip("randint"),
+    skip("randn"),
+    skip("scalar_tensor"),
+    skip("signal.windows.bartlett"),
+    skip("signal.windows.blackman"),
+    skip("signal.windows.cosine"),
+    skip("signal.windows.exponential"),
+    skip("signal.windows.gaussian"),
+    skip("signal.windows.general_cosine"),
+    skip("signal.windows.general_hamming"),
+    skip("signal.windows.hamming"),
+    skip("signal.windows.hann"),
+    skip("signal.windows.kaiser"),
+    skip("signal.windows.nuttall"),
+    skip("sparse.sampled_addmm"),
+    skip("squeeze"),
+    skip("zeros"),
+}
+
+
+# =============================================================================
+# Base tensor unbacked (for test_ops_unbacked.py, not DTensor)
+# =============================================================================
+unbacked_base_xfails = {
+    xfail("_chunk_cat"),
+    xfail("_unsafe_masked_index_put_accumulate"),
+    xfail("_upsample_bilinear2d_aa"),
+    xfail("addmv"),
+    xfail("allclose"),
+    xfail("as_strided_scatter"),
+    xfail("baddbmm"),
+    xfail("bernoulli"),
+    xfail("cauchy"),
+    xfail("cdist"),
+    xfail("cholesky"),
+    xfail("chunk"),
+    xfail("combinations"),
+    xfail("corrcoef"),
+    xfail("cov"),
+    xfail("cross"),
+    xfail("cummax"),
+    xfail("cummin"),
+    xfail("cumulative_trapezoid"),
+    xfail("diagonal_scatter"),
+    xfail("diff"),
+    xfail("dist"),
+    xfail("dsplit"),
+    xfail("equal"),
+    xfail("exponential"),
+    xfail("fft.fft"),
+    xfail("fft.fft2"),
+    xfail("fft.fftn"),
+    xfail("fft.fftshift"),
+    xfail("fft.hfft"),
+    xfail("fft.hfft2"),
+    xfail("fft.hfftn"),
+    xfail("fft.ifft"),
+    xfail("fft.ifft2"),
+    xfail("fft.ifftn"),
+    xfail("fft.ifftshift"),
+    xfail("fft.ihfft"),
+    xfail("fft.ihfft2"),
+    xfail("fft.ihfftn"),
+    xfail("fft.irfft"),
+    xfail("fft.irfft2"),
+    xfail("fft.irfftn"),
+    xfail("fft.rfft"),
+    xfail("fft.rfft2"),
+    xfail("fft.rfftn"),
+    xfail("float"),
+    xfail("geometric"),
+    xfail("geqrf"),
+    xfail("gradient"),
+    xfail("grid_sampler_2d"),
+    xfail("hash_tensor"),
+    xfail("histogram"),
+    xfail("histogramdd"),
+    xfail("hsplit"),
+    xfail("index_fill"),
+    xfail("inner"),
+    xfail("kron"),
+    xfail("linalg.cond"),
+    xfail("linalg.cross"),
+    xfail("linalg.householder_product"),
+    xfail("linalg.ldl_solve"),
+    xfail("linalg.lstsq"),
+    xfail("linalg.lstsq", "grad_oriented"),
+    xfail("linalg.lu_solve"),
+    xfail("linalg.matrix_norm"),
+    xfail("linalg.matrix_power"),
+    xfail("linalg.matrix_rank"),
+    xfail("linalg.matrix_rank", "hermitian"),
+    xfail("linalg.multi_dot"),
+    xfail("linalg.norm"),
+    xfail("linalg.norm", "subgradients_at_zero"),
+    xfail("linalg.pinv"),
+    xfail("linalg.pinv", "hermitian"),
+    xfail("linalg.pinv", "singular"),
+    xfail("linalg.qr"),
+    xfail("linalg.solve"),
+    xfail("linalg.solve_ex"),
+    xfail("linalg.solve_triangular"),
+    xfail("linalg.svd"),
+    xfail("linalg.svdvals"),
+    xfail("linalg.tensorinv"),
+    xfail("linalg.tensorsolve"),
+    xfail("linalg.vander"),
+    xfail("linalg.vector_norm"),
+    xfail("log_normal"),
+    xfail("logdet"),
+    xfail("logsumexp"),
+    xfail("lu_solve"),
+    xfail("lu_unpack"),
+    xfail("masked.amax"),
+    xfail("masked.amin"),
+    xfail("masked.argmax"),
+    xfail("masked.argmin"),
+    xfail("masked.cumprod"),
+    xfail("masked.cumsum"),
+    xfail("masked.log_softmax"),
+    xfail("masked.logaddexp"),
+    xfail("masked.logsumexp"),
+    xfail("masked.mean"),
+    xfail("masked.median"),
+    xfail("masked.norm"),
+    xfail("masked.prod"),
+    xfail("masked.softmax"),
+    xfail("masked.softmin"),
+    xfail("masked.std"),
+    xfail("masked.sum"),
+    xfail("masked.var"),
+    xfail("max_pool2d_with_indices_backward"),
+    xfail("multinomial"),
+    xfail("nanquantile"),
+    xfail("nn.functional.adaptive_avg_pool1d"),
+    xfail("nn.functional.adaptive_avg_pool2d"),
+    xfail("nn.functional.adaptive_avg_pool3d"),
+    xfail("nn.functional.adaptive_max_pool1d"),
+    xfail("nn.functional.adaptive_max_pool2d"),
+    xfail("nn.functional.adaptive_max_pool3d"),
+    xfail("nn.functional.alpha_dropout"),
+    xfail("nn.functional.avg_pool1d"),
+    xfail("nn.functional.avg_pool2d"),
+    xfail("nn.functional.avg_pool3d"),
+    xfail("nn.functional.batch_norm"),
+    xfail("nn.functional.bilinear"),
+    xfail("nn.functional.binary_cross_entropy"),
+    xfail("nn.functional.binary_cross_entropy_with_logits"),
+    xfail("nn.functional.channel_shuffle"),
+    xfail("nn.functional.conv1d"),
+    xfail("nn.functional.conv2d"),
+    xfail("nn.functional.conv3d"),
+    xfail("nn.functional.conv_transpose1d"),
+    xfail("nn.functional.conv_transpose2d"),
+    xfail("nn.functional.conv_transpose3d"),
+    xfail("nn.functional.cosine_similarity"),
+    xfail("nn.functional.cross_entropy"),
+    xfail("nn.functional.ctc_loss"),
+    xfail("nn.functional.dropout"),
+    xfail("nn.functional.dropout2d"),
+    xfail("nn.functional.dropout3d"),
+    xfail("nn.functional.embedding"),
+    xfail("nn.functional.embedding_bag"),
+    xfail("nn.functional.feature_alpha_dropout", "with_train"),
+    xfail("nn.functional.feature_alpha_dropout", "without_train"),
+    xfail("nn.functional.fractional_max_pool2d"),
+    xfail("nn.functional.fractional_max_pool3d"),
+    xfail("nn.functional.gaussian_nll_loss"),
+    xfail("nn.functional.grid_sample"),
+    xfail("nn.functional.group_norm"),
+    xfail("nn.functional.huber_loss"),
+    xfail("nn.functional.instance_norm"),
+    xfail("nn.functional.interpolate", "area"),
+    xfail("nn.functional.interpolate", "bicubic"),
+    xfail("nn.functional.interpolate", "bilinear"),
+    xfail("nn.functional.interpolate", "linear"),
+    xfail("nn.functional.interpolate", "trilinear"),
+    xfail("nn.functional.l1_loss"),
+    xfail("nn.functional.local_response_norm"),
+    xfail("nn.functional.max_pool1d"),
+    xfail("nn.functional.max_pool2d"),
+    xfail("nn.functional.max_pool3d"),
+    xfail("nn.functional.max_unpool1d"),
+    xfail("nn.functional.max_unpool1d", "grad"),
+    xfail("nn.functional.max_unpool2d"),
+    xfail("nn.functional.max_unpool2d", "grad"),
+    xfail("nn.functional.max_unpool3d"),
+    xfail("nn.functional.max_unpool3d", "grad"),
+    xfail("nn.functional.mse_loss"),
+    xfail("nn.functional.multi_head_attention_forward"),
+    xfail("nn.functional.multilabel_margin_loss"),
+    xfail("nn.functional.nll_loss"),
+    xfail("nn.functional.pad", "circular"),
+    xfail("nn.functional.pad", "reflect"),
+    xfail("nn.functional.pad", "replicate"),
+    xfail("nn.functional.pad", "replicate_negative"),
+    xfail("nn.functional.pdist"),
+    xfail("nn.functional.pixel_shuffle"),
+    xfail("nn.functional.prelu"),
+    xfail("nn.functional.rrelu"),
+    xfail("nn.functional.scaled_dot_product_attention"),
+    xfail("nn.functional.smooth_l1_loss"),
+    xfail("nn.functional.unfold"),
+    xfail("nn.functional.upsample_bilinear"),
+    xfail("norm"),
+    xfail("norm", "fro"),
+    xfail("norm", "nuc"),
+    xfail("normal"),
+    xfail("normal", "in_place"),
+    xfail("normal", "number_mean"),
+    xfail("ormqr"),
+    xfail("pca_lowrank"),
+    xfail("pinverse"),
+    xfail("quantile"),
+    xfail("qr"),
+    xfail("rand_like"),
+    xfail("randint_like"),
+    xfail("randn_like"),
+    xfail("repeat_interleave"),
+    xfail("resize_"),
+    xfail("resize_as_"),
+    xfail("roll"),
+    xfail("scatter"),
+    xfail("scatter_add"),
+    xfail("scatter_reduce", "amax"),
+    xfail("scatter_reduce", "amin"),
+    xfail("scatter_reduce", "mean"),
+    xfail("scatter_reduce", "prod"),
+    xfail("scatter_reduce", "sum"),
+    xfail("searchsorted"),
+    xfail("sparse.mm", "reduce"),
+    xfail("split"),
+    xfail("stft"),
+    xfail("svd"),
+    xfail("svd_lowrank"),
+    xfail("sum_to_size"),
+    xfail("take"),
+    xfail("take_along_dim"),
+    xfail("tensordot"),
+    xfail("tensor_split"),
+    xfail("to_sparse"),
+    xfail("trapezoid"),
+    xfail("trapz"),
+    xfail("unbind"),
+    xfail("unbind_copy"),
+    xfail("uniform"),
+    xfail("unsafe_chunk"),
+    xfail("unsafe_split"),
+    xfail("view_as_complex"),
+    xfail("vsplit"),
+    # skips
+    skip("arange"),
+    skip("broadcast_shapes"),
+    skip("empty"),
+    skip("empty_permuted"),
+    skip("empty_strided"),
+    skip("eye"),
+    skip("full"),
+    skip("item"),
+    skip("linspace"),
+    skip("linspace", "tensor_overload"),
+    skip("logspace"),
+    skip("logspace", "tensor_overload"),
+    skip("ones"),
+    skip("randint"),
+    skip("randn"),
+    skip("scalar_tensor"),
+    skip("signal.windows.bartlett"),
+    skip("signal.windows.blackman"),
+    skip("signal.windows.cosine"),
+    skip("signal.windows.exponential"),
+    skip("signal.windows.gaussian"),
+    skip("signal.windows.general_cosine"),
+    skip("signal.windows.general_hamming"),
+    skip("signal.windows.hamming"),
+    skip("signal.windows.hann"),
+    skip("signal.windows.kaiser"),
+    skip("signal.windows.nuttall"),
+    skip("zeros"),
+    skip("sparse.sampled_addmm"),
+}
